Adding an example using gorilla mux and direct http control.
