--- cmd/samples/httpb/receiver/main.go
@@ -1,29 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	cloudevents "github.com/cloudevents/sdk-go"
-)
-
-func main() {
-	ctx := context.Background()
-	p, err := cloudevents.NewHTTP()
-	if err != nil {
-		log.Fatalf("failed to create protocol: %s", err.Error())
-	}
-
-	c, err := cloudevents.NewClient(p)
-	if err != nil {
-		log.Fatalf("failed to create client, %v", err)
-	}
-
-	log.Printf("will listen on :8080\n")
-	log.Fatalf("failed to start receiver: %s", c.StartReceiver(ctx, receive))
-}
-
-func receive(ctx context.Context, event cloudevents.Event) {
-	fmt.Printf("%s", event)
-}
